Consolidate those test cases
