+++ Tests/test_dict.py
@@ -82,6 +82,21 @@ def test_dict_merging():
     a |= b
     assert a == {'x': 'y'}
 
+
+@pytest.mark.optimization(1)
+def test_dict_keys_optimization():
+    def _f():
+        l = {'a': 1, 'b': 2}
+        k = l.keys()
+        return tuple(k)
+
+    assert _f() == ('a', 'b')
+    inf = pyjion.info(_f)
+    assert inf.compiled
+    assert inf.optimizations & pyjion.OptimizationFlags.BuiltinMethods
+
+
+
 def test_large_const_dict():
     ANSI_COLOR_NAMES = {
         "black": 0,
